chore(test): move ref tests to new file
